# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/messengerRelatedUsers": {
            "get": {
                "description": "Retrieves a messenger-related user by chatID, messengerUserID, userID and messengerID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Get messenger-related user by chatID, messengerUserID, userID and messengerID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chat_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Messenger User ID",
                        "name": "messenger_user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Messenger ID",
                        "name": "messenger_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messenger-related user details",
                        "schema": {
                            "$ref": "#/definitions/dto.MessengerRelatedUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Messenger-related user not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new messenger-related user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Create a new messenger-related user",
                "parameters": [
                    {
                        "description": "MessengerRelatedUser to create",
                        "name": "messenger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMessengerRelatedUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created messenger-related user ID",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/messengerRelatedUsers/all": {
            "get": {
                "description": "Retrieves all messenger-related users with pagination and ordering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Get all messenger-related users",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size (default: 50)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at DESC",
                        "description": "Order by field (default: created_at DESC)",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of messenger-related users",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedMessengerRelatedUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/messengerRelatedUsers/{messenger_user_id}/user": {
            "get": {
                "description": "Retrieves a userID by messengerUserID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Get a userID by messengerUserID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Messenger UserID",
                        "name": "messenger_user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User ID",
                        "schema": {
                            "$ref": "#/definitions/dto.UserIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/messengers": {
            "get": {
                "description": "Retrieves all messengers with pagination and ordering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Get all messengers",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size (default: 50)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at DESC",
                        "description": "Order by field (default: created_at DESC)",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of messengers",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedMessengersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new messenger type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Create a new messenger type",
                "parameters": [
                    {
                        "description": "Messenger to create",
                        "name": "messenger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMessengerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created messenger ID",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/messengers/by-name/{messenger_name}": {
            "get": {
                "description": "Retrieves a messenger ID by its name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Get messenger ID by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Messenger name",
                        "name": "messenger_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messenger ID",
                        "schema": {
                            "$ref": "#/definitions/dto.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Messenger not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/messengers/{messenger_id}": {
            "get": {
                "description": "Retrieves a messenger by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messengers"
                ],
                "summary": "Get messenger by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Messenger ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messenger details",
                        "schema": {
                            "$ref": "#/definitions/dto.MessengerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Messenger not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateMessengerRelatedUserRequest": {
            "type": "object",
            "required": [
                "chat_id",
                "messenger_id",
                "messenger_user_id",
                "user_id"
            ],
            "properties": {
                "chat_id": {
                    "type": "string",
                    "example": "-1001234567890"
                },
                "messenger_id": {
                    "type": "integer",
                    "example": 1
                },
                "messenger_user_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.CreateMessengerRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "telegram"
                }
            }
        },
        "dto.CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
        "dto.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.MessengerRelatedUserResponse": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string",
                    "example": "-1001234567890"
                },
                "created_at": {
                    "description": "ISO 8601 format",
                    "type": "string",
                    "example": "2024-01-10T08:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "messenger_id": {
                    "type": "integer",
                    "example": 1
                },
                "messenger_user_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "updated_at": {
                    "description": "ISO 8601 format",
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.MessengerResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "ISO 8601 format",
                    "type": "string",
                    "example": "2024-01-10T08:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "telegram"
                }
            }
        },
...
        "dto.PaginatedMessengerRelatedUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MessengerRelatedUserResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                }
            }
        },
        "dto.PaginatedMessengersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MessengerResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                }
            }
        },
...
        "dto.UserIDResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}